+++ src/core/wallets/operations/taproot-utils/noble_ecc.ts
